Fix ordering of scripts for Persona
The Persona login JS needs to load before the MediaWiki
hooks do, otherwise the navigator API will not be available.

Unfortunately, MediaWiki throws all non-Resource Loader script
files at the end of all scripts, so the only way to make sure
it is loaded before is to put is as a head item.

Bug: 57525
Change-Id: I86e5b2484aaae2d3001dc03c333be7d32ee078f2